Removed some fontello stuff that was causing a malfunction.
The social icons were doubled because the entypo fontello thing was adding a second set of social icons in addition to the SVGs embedded in the footer.  This commit removes the stuff that triggers the fontello thing.  This footer is embedded in sites where we can’t turn that fontello thing off, for the main page.